Modifiche User Media Post e Refactoring
