--- src/main/java/com/github/terravivaproject/terraviva/entities/Variety.java
@@ -1,30 +1,0 @@
-package com.github.terravivaproject.terraviva.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Entity
-@Table
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Variety {
-    @Id
-    @NotBlank
-    @NotNull
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotNull
-    @NotBlank
-    @Column(unique = true)
-    //TODO add index on name
-    private String name;
-}
