+++ src/main/java/com/github/terravivaproject/terraviva/media/entities/Media.java
@@ -1,0 +1,47 @@
+package com.github.terravivaproject.terraviva.media.entities;
+
+import com.github.terravivaproject.terraviva.user.entities.User;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.Accessors;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Entity
+@Table
+@Getter
+@Setter
+@Accessors(chain = true)
+@NoArgsConstructor
+@AllArgsConstructor
+public class Media {
+
+    @Id
+    @NotBlank
+    @NotNull
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private UUID id;
+
+    @NotBlank @NotNull
+    private LocalDateTime createdOn = LocalDateTime.now();
+
+    @NotBlank
+    @NotNull
+    @Column(length = 50)
+    private String location;
+
+    @NotBlank
+    @NotNull
+    @Column(length = 32)
+    private String mediaMd5;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User owner;
+
+}
