+++ src/main/java/com/github/terravivaproject/terraviva/seed/entities/Variety.java
@@ -1,45 +1,30 @@
-package com.github.terravivaproject.terraviva.entities;
+package com.github.terravivaproject.terraviva.seed.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.experimental.Accessors;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 
 @Entity
 @Table
 @Getter
 @Setter
-@Accessors(chain = true)
 @NoArgsConstructor
 @AllArgsConstructor
-public class Varieties {
-
+public class Variety {
     @Id
     @NotBlank
     @NotNull
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-
-    @NotBlank
     @NotNull
+    @NotBlank
+    @Column(unique = true)
+    //TODO add index on name
     private String name;
-
-
-    @OneToOne
-    private Species species;
-
-    //TODO implements @Pattern -> URI
-    private String variety_image;
-
-    @NotBlank @NotNull
-    private LocalDateTime dateTime = LocalDateTime.now();
-
-
 }
