+++ src/main/java/com/github/terravivaproject/terraviva/social/controllers/PostController.java
@@ -1,0 +1,4 @@
+package com.github.terravivaproject.terraviva.social.controllers;
+
+public class PostController {
+}
